Refactored datetime related files into their own directory
