+++ frontend/src/components/Layout/Layout.jsx
@@ -7,8 +7,8 @@ import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import SidebarContent from "./SidebarContent";
 import Mapbox from "../Map/Mapbox";
-import RangeFactory from "./Range";
-import { parseTimePeriods } from "./DateTimeParser";
+import RangeFactory from "./Datetime/Range";
+import { parseTimePeriods } from "./Datetime/DateTimeParser";
 import {getLinksBetweenNodes, getTravelDataFile} from "../../actions/actions";
 import "./Layout.css";
 import FormControl from "@material-ui/core/FormControl";
