+++ frontend/src/components/Layout/SidebarContent.jsx
@@ -10,8 +10,8 @@ import FormLabel from "@material-ui/core/FormLabel";
 import FormControl from "@material-ui/core/FormControl";
 import FormGroup from "@material-ui/core/FormGroup";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
-import { MAX_DATE, MIN_DATE } from "./Range";
-import Presets from "./Presets"
+import { MAX_DATE, MIN_DATE } from "./Datetime/Range";
+import Presets from "./Datetime/Presets"
 import Days from "./Days";
